Update helper.rb
Using a helper file like this makes this app stand out from the crowd! Well done. 

The only thing I'd do is maybe add a few comments at the start of the file explaining what you are doing in this file and may be a break down of the methods and their corresponding line number.That way if you were to had this to another dev they could quickly find methods.